requirements.txt
"""
Requirements for Cultural Bias Shield
"""
Flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
